Add InTune support
* Microsoft InTune is a device management/compliance checker
that some organizations use for accessing corporate resources.
When using an InTune enrolled device, Edge talks to the identity
service using dbus for seamless login to corporate resources, and
for conditional access enforcement. In order to support this, the
msalsdk-dbusclient shared library must be present. This commit adds
this library and its dependency (sd-bus) to the flatpak

* NOTE: This won't work until properly flatpak/xdg-dbus-proxy#21 is fixed